Move make_dict_observable into own file
